adapting luke's requested changes.
