+++ docs/fundamentals/backends.rst
@@ -123,25 +123,24 @@ There is no need to install QiliSim separately, as it is included with the core 
 **Parameters**
 
 - ``noise_model`` (:class:`~qilisdk.noise.noise_model.NoiseModel`, optional): Noise model applied during simulation.
-- ``analog_simulation_method`` (:class:`~qilisdk.backends.qilisim.AnalogMethod`, optional): Analog simulation method and method-specific options.
-- ``digital_simulation_method`` (:class:`~qilisdk.backends.qilisim.DigitalMethod`, optional): Digital simulation options.
-- ``execution_config`` (:class:`~qilisdk.backends.qilisim.ExecutionConfig`, optional): Runtime execution options such as thread count and random seed.
+- ``analog_simulation_method`` (:class:`~qilisdk.backends.backend_config.AnalogMethod`, optional): Analog simulation method and method-specific options.
+- ``digital_simulation_method`` (:class:`~qilisdk.backends.backend_config.DigitalMethod`, optional): Digital simulation options.
+- ``execution_config`` (:class:`~qilisdk.backends.backend_config.ExecutionConfig`, optional): Runtime execution options such as thread count and random seed.
 
 **Configuration example**
 
 .. code-block:: python
 
     from qilisdk.backends import QiliSim
-    from qilisdk.backends.qilisim import AnalogMethod, DigitalMethod, ExecutionConfig, MonteCarloConfig
+    from qilisdk.backends.backend_config import AnalogMethod, DigitalMethod, ExecutionConfig, MonteCarloConfig
 
     backend = QiliSim(
         analog_simulation_method=AnalogMethod.arnoldi(
             dim=16,
-            num_substeps=2,
-            monte_carlo=MonteCarloConfig(trajectories=200),
+            num_substeps=2
         ),
         digital_simulation_method=DigitalMethod.state_vector(max_cache_size=2_000),
-        execution_config=ExecutionConfig(num_threads=4, seed=42),
+        execution_config=ExecutionConfig(num_threads=4, seed=42, monte_carlo=MonteCarloConfig(trajectories=200)),
     )
 
 **Example**
